refactor members logic to slack service
